Updates for Save Progress, Toggle Arrows, and Bulk Editing
resolve #28
resolve #27
affect #3

+++ Assets/Scripts/Controllers/ResourceController.cs
@@ -84,7 +84,7 @@ public class ResourceController : MonoBehaviour
         }
     }
 
-    public void SetResource(Resource resource, int quantity)
+    public void SetResource(Resource resource, long quantity)
     {
         this.resource = new Resource(resource);
         this.resource.quantity = quantity;
